+++ cookbooks/aws-parallelcluster-install/recipes/slurm.rb
@@ -20,6 +20,7 @@ return if node['conditions']['ami_bootstrapped']
 include_recipe 'aws-parallelcluster-install::base'
 include_recipe 'aws-parallelcluster-install::munge'
 include_recipe 'aws-parallelcluster-install::pmix'
+include_recipe "aws-parallelcluster-install::intel"
 
 package %w[slurm* libslurm*] do
   action :purge